impr: Remove not needed log for logging (#3934)
Every log message needs to acquire a lock to evaluate whether the logger
should log it. #3303 added this logic to fix a data race the thread
sanitizer job found. As the SDK only calls the configure method when
starting, we don't need to put a synchronized keyword around the code
that evaluates the log level. This PR replaces the synchronized keyword
by ignoring the thread sanitizer.