fix: Crash in baggageEncodedDictionary (#4017)
The baggageEncodedDictionary method sometimes crashes with a null
pointer exception in CFCharacterSetAddCharactersInString. It seems
mutableCopy of NSCharacterSet.alphanumericCharacterSet or when calling
stringByAddingPercentEncodingWithAllowedCharacters sometimes doesn't
work. Instead of trying to fix the underlying issue of a null pointer
exception, we can convert the code to Swift, designed for safety to
avoid such mistakes. This should prevent the crash.


Co-authored-by: Dhiogo Brustolin <dhiogo.brustolin@sentry.io>