fix: Potential deadlock in app hang detection (#4063)
The app hang detection retrieves the stacktraces of all threads when it
detects an app hang. The logic for retrieving the stacktrace uses
sentrycrashdl_dladdr, which isn't async-safe, although it claims to be
because it accesses _dyld_get_image_header, which acquires a lock.
Therefore, in rare cases, retrieving the stacktrace for all threads can
lead to a deadlock. We fix this by skipping symbolication, which calls
sentrycrashdl_dladdr when debug is false, which we we thought we did in

Fixes GH-4056