fix: Missing mach info for crash reports
The SentryInternalCDefines didn't define the SENTRY_HOST_APPLE, which
led to missing mach info for crash reports, as the SentryCrashReport
relies on the SENTRY_HOST_APPLE to write that info. For reference, the
SentryCrashCRASH_HOST_APPLE was changed to SENTRY_HOST_APPLE in GH-4101.
The problem is fixed now by defining SENTRY_HOST_APPLE in
SentryInternalCDefines.

Fixes GH-4227