fix: Rare flush timeout when called in tight loop
Fix an edge case that flush could keep timing out when called in a tight
loop. The test case testFlush_WhenNoInternet_BlocksAndFinishes does this
and sometimes fails with 'Flush should not time out'. This is fixed now,
by removing the duplicate setting of isFlushing to NO in two different
synchronize blocks.