ci: Remove unit tests Xcode 14 for tvOS and macOS (#4256)
Both tvOS and macOS were using Xcode 14.3 on macos-13, which hasn't been
available for a while now. So, the tests were actually running on Xcode
15.2, which uses tvOS 17.2 instead of tvOS 16 and macOS 14.2 instead of
macOS 13. We never noticed this. Furthermore, I think it's not worth the
computation time to run tests on these OS versions compared to the small
risk of finding bugs. We run the tests on tvOS 15 and macOS 12, which
should be sufficient.