ci: Remove UI Address Sanitizer Tests (#4259)
The tests keep failing in CI for no good reason. We don't act on the
failures, so it's better to disable them. It's better to have a test
suite that we can trust so we can take a step back when a CI job fails
instead of continuing to ignore these failures.