fix: Guard dereferencing of stack frame pointer in SentryBacktrace
This occasionally crashes when trying to read the stack frame. The stack bounds check is insufficient to catch this case, so we can try adding a guard when reading the memory. We may still get garbage results for `returnAddress`, but that would just result in an unknown symbol at symbolication time rather than a runtime crash.