test: Switch testCorruptEnvelope to a unit test (#4348)
Change the testCorruptEnvelope to a unit test instead of using a UITest.
Tapping the corruptEnvelope button after v 8.33.0 doesn't even create an
envelope because the writing of the envelope is atomic. The app
terminates without writing an envelope. As the UITest was flaky, we
instead switched to a unit test validating the underlying problem of
deserializing a corrupted envelope. Furthermore, we have to ensure that
the corrupted envelopes don't crash when deserializing them because we
have no guarantee that our users don't modify the envelope files.