impr: Slightly speed up SentryInAppLogic (#4370)
Slightly speed up the SentryInAppLogic by avoiding calls to
lastPathComponent and lowercaseString in loops. Both properties require
some extra work, such as iterating over the contents of the string. This
is noticeable for larger apps with many images and inAppIncludes and
inAppExcludes. For small apps, this doesn't make a difference.