fix: Edge case for swizzleClassNameExclude
Skip creating transactions for UIViewControllers ignored for swizzling
via the option swizzleClassNameExclude. Due to some edge cases with nib
files, the SDK doesn't swizzle the loadView method of the
UIViewController subclasses, but instead, it swizzles the
UIViewController.loadView method directly. Although the SDK doesn't
swizzle the classes specified in swizzleClassNameExclude, it created
transactions. Now, this is fixed.

Fixes GH-4386