fix: Make SentryDebugImageProvider internal
Move SentryDebugImageProvider.h to HybridPublic, as it's not supposed to
be public. Strictly speaking, this is a breaking change, but nobody
should have to use this API except the hybrid SDKs. We can always make
it public again if we get complaints. Furthermore, remove a method not
used by the SDK or hybrid SDKs.

Fixes GH-3056