fix: Finish TTID when viewWillAppear is skipped (#4417)
According to the Apple docs, viewWillAppear should be called before the
UIViewController is added to the view hierarchy. There are some edge
cases, though, when this doesn't happen, and we saw customers'
transactions also proofing this. Therefore, we must also report the
initial display here, as the customers' transactions had spans for
viewWillLayoutSubViews.

Fixes GH-4383