fix: Missing UIViewController traces with Xcode 16 (#4523)
Xcode 16 introduces a new flag ENABLE_DEBUG_DYLIB. If this flag is
enabled, debug builds of app and app extension targets on supported
platforms and SDKs will be built with the main binary code in a separate
“NAME.debug.dylib”. Our swizzling logic for UIViewControllers sometimes
skipped swizzling the UIViewControllers in that debug.dylib binary. This
is fixed now, by swizzling all matching inAppInclude binary images that
the binary image cache returns and not only the first one.