fix: Remove internal unknown dict for Breadcrumbs (#4803)
Remove unknown dict when initializing the crumbs with a dict added in
#2820. The Java SDK, for example, implements this pattern to keep
unknown JSON properties when deserializing and serializing classes. This
allows more flexibility when a hybrid SDK or the native SDK serializes
an event to disk, and the Java SDK deserializes this event. With the
unknown dict, the Java SDK doesn't need to implement all the properties
other SDKs have. We only have this for crumbs on the Cocoa SDK, and it
doesn't make sense only to keep it for one class. Instead, we removed it
to simplify the code because we would need to implement some complex
deserialization logic with Swift Decodable for deserializing events. We
can always reconsider this approach, but then we would need to add the
unknown property concept for all serializable/decodable classes.