fix: Correctly finish TTFD span when no new frame
This PR fixes wrongly finishing the TTFD span as the deadline exceeded
when the user calls finishSpansIfNotFinished, and the
UIViewControllerPerformanceTracker calls finishSpansIfNotFinished before
a new frame gets drawn. This happened for a customer using a
SplashScreenUIViewController. Now, the TimeToDisplayTracker finishes the
TTFD span in finishSpansIfNotFinished when the user called
finishSpansIfNotFinished.