fix: Crash in setMeasurement when name is nil (#5064)
Although the name for setMeasurement is nonnull, we saw occurrences when
it was nil, which led to crashes. Now, when the name is nil, the SDK
discards the measurement and logs an error.

Calling setMeasurement on transactions ans spans could lead to crashes.
This is fixed now by making it thread-safe.