ref: Fix compiler errors for async safe logs (#5207)
When setting the SENTRY_ASYNC_SAFE_LOG_LEVEL to
SENTRY_ASYNC_SAFE_LOG_DEBUG, the project fails to compile because
SentrySessionReplaySyncC contains compiler errors for the log levels.
This is fixed now, and this PR also adds a check in CI to prevent this
from happening again, by setting the SENTRY_ASYNC_SAFE_LOG_LEVEL to
SENTRY_ASYNC_SAFE_LOG_TRACE and compiling the project.