fix: Missing debug meta for non fatal events (#5352)
GH-5225 introduced a bug so that nonfatal events have missing
debug_meta, leading to redacted stacktraces. Fatal events such as
crashes aren't impacted by this bug. The root cause was that the
initializers of specific classes were getting initialized in the
SentryDependencyContainer, which also accessed the shared instance of
the SentryDependencyContainer, causing a chicken-egg problem. This is
fixed now by removing the usage of SentryDependencyContainer of the
dependencies initialized in SentryDependencyContainer.init. Furthermore,
the SentryDependencyContainer now uses SENTRY_GRACEFUL_FATAL to exit the
application in debug and log a warning in production.