ci: Use xcrun for crash UI test (#5512)
XCTest isn't built for crashing during tests. Instead of using XCTest to press a button and let a test app crash, we now use UserDefaults to tell the test app to crash during launch. We then simply launch the app again and wait to see if it keeps running. This is basically the same as the testCrash of the SwiftUITestSample without using XCTests.