test: Improve corrupted envelope UI test (#5549)
Simplify the UI test by writing the corrupted envelope and then starting
the SDK. This removes the need to relaunch the app, which is often
problematic in CI. Furthermore, this PR now lets the UI test fail with
an error message if the test can't write the corrupted envelope instead
of using fatalError. So when this fails, we know clearly what went
wrong.