fix: Memory leak in SentrySessionReplayIntegration (#5770)
The global event processor used a strong reference to self which lead to
strong reference cycle, meaning that the ARC couldn't dealloc
SentrySessionReplayIntegration. This also led to having multiple
SentrySessionReplayIntegrations running in the tests after running the
SentrySessionReplayIntegrationTests, making the tests flaky. This is
fixed now by using a weak reference.