test: Use XCTestExpectation for concurrent mods (#5850)
Replace the dispatch group with an XCTestExpectation to get a failure
message when the wait times out. To make this work this PR changes the
method to an extension so it can call XCTestCase.wait. Furthermore, this
PR reduces the timeout from 500s to 10s, because 500s is way to high
and risks slowing down CI.

Contributes to GH-5789