ref: Remove unused session.user (#6897)
The session.user isn't really used anywhere in the code.
While the scope applied the user to the session, the session
doesn't do anything with the user property. Therefore, we can remove it.
Also, the develop docs don't mention a user property. Only the
distinctID that should be the userID.